api/requirements.txt
..Flask==2.2.3
Werkzeug==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
Flask-Swagger-UI==4.11.1
Flask-Cors==3.1.1
Flasgger==0.9.7.1
